[Feat] chapter 2-11 URL 프리픽스 알아 두기

+++ src/main/java/com/example/question/service/QuestionService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.example.DataNotFoundException;
 import com.example.question.model.Question;
 import com.example.question.repository.QuestionRepository;
 
@@ -20,7 +21,7 @@ public class QuestionService {
         return this.questionRepository.findAll();
     }
     
-    public Question getQuestion(Integer id) {  
+    public Question getQuestion(Long id) {  
         Optional<Question> question = this.questionRepository.findById(id);
         if (question.isPresent()) {
             return question.get();
